apps/fastapi/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiomultiprocess==0.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
arrow==0.15.2
async-timeout==5.0.1
attrs==24.2.0
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cursor==1.3.5
daily-python==0.13.0
dataclasses-json==0.6.7
deepgram-sdk==3.7.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
frozenlist==1.5.0
google-api-python-client==1.7.11
google-auth==2.36.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
h11==0.14.0
halo==0.0.28
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
isort==5.13.2
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
log-symbols==0.0.14
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
openai==1.55.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prisma==0.15.0
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyparsing==3.2.0
python-dateutil==2.8.0
python-dotenv==1.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
rsa==4.9
six==1.16.0
slackclient==2.2.1
sniffio==1.3.1
spinners==0.0.24
starlette==0.41.2
StrEnum==0.4.15
tenacity==9.0.0
termcolor==2.5.0
tomli==2.1.0
tomlkit==0.13.2
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==3.0.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==14.1
yarl==1.17.2
